requirements.txt
fastapi 
uvicorn 
python-dotenv 
langchain-google-genai 
langchain-pinecone 
langchain-core 
tiktoken 
pinecone-client 
cohere